+++ src/OrthoNormalBasis.jl
@@ -14,8 +14,8 @@ function create_onb_from_z(normal::Union{Vec, Normal})
     a = -1.0/(s + normal.z)
     b = normal.x * normal.y * a
 
-    e1 = Vec(1.0 + s * normal.x^2 * a, s * b, -s * normalx)
-    e2 = Vec(b, sign + normal.y * normal.y * a, -normal.y)
+    e1 = Normal(1.0 + s * normal.x^2 * a, s * b, -s * normal.x)
+    e2 = Normal(b, sign + normal.y^2 * a, -normal.y)
 
-    return e1, e2, Vec(normal.x, normal.y, normal.z)
+    return e1, e2, Normal(normal.x, normal.y, normal.z)
 end